requirements.txt
amqp==2.5.2
apipkg==1.5
argh==0.26.2
arrow==0.15.5
asgiref==3.2.3
asttokens==2.0.3
attrs==19.3.0
autobahn==20.1.3
Automat==0.8.0
base58==2.0.0
billiard==3.6.3.0
bitarray==1.2.1
celery==4.4.1
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
crc16==0.1.1
cryptography==2.8
cytoolz==0.10.1
daphne==2.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-archive==0.1.6
django-dbbackup==3.3.0
django-dramatiq==0.9.1
django-filter==2.2.0
django-import-export==2.0.2
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
dramatiq==1.8.1
ed25519==1.5
et-xmlfile==1.0.1
eth-abi==2.1.1
eth-account==0.5.2
eth-brownie==1.8.4
eth-event==1.0.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
execnet==1.7.1
Flask==1.1.1
Flask-Web3==0.1.1
funcsigs==1.0.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
hexbytes==0.2.0
hyperlink==19.0.0
hypothesis==5.11.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
inflection==0.4.0
ipfshttpclient==0.4.13.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jsonschema==3.2.0
kombu==4.6.8
lru-dict==1.1.6
MarkupPy==1.14
MarkupSafe==1.1.1
mnemonic==0.19
more-itertools==8.2.0
multiaddr==0.0.9
mythx-models==1.7.5
netaddr==0.7.19
numpy==1.18.1
odfpy==1.4.1
opencage==1.2
openpyxl==3.0.3
packaging==20.3
parsimonious==0.8.1
pathtools==0.1.2
pbkdf2==1.3
pika==1.1.0
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
psycopg2==2.8.4
pure-interface==4.0.0
py==1.8.1
py-solc==3.2.0
py-solc-ast==1.2.4
py-solc-x==0.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.7
Pygments==2.6.1
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.2
pytest-forked==1.1.3
pytest-xdist==1.32.0
python-dateutil==2.8.1
python-decouple==3.3
python-interface==1.5.3
pythx==1.5.7
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.23.0
rlp==1.2.0
semantic-version==2.8.5
service-identity==18.1.0
six==1.14.0
solc==0.0.0a0
sortedcontainers==2.1.0
sqlparse==0.3.0
sseclient==0.0.24
stellar-base==1.1.2.0
tablib==1.1.0
toolz==0.10.0
tqdm==4.46.0
twilio==6.21.0
Twisted==19.10.0
txaio==20.1.1
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
varint==1.0.2
vine==1.3.0
vyper==0.1.0b17
watchdog==0.8.3
watchdog-gevent==0.1.0
wcwidth==0.1.9
web3==5.9.0
websockets==8.1
Werkzeug==1.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==4.7.1
